changed deleteThought to action-creator
moved rawTextToHTML to importText file
