+++ src/action-creators/index.js
@@ -7,11 +7,14 @@ export { default as cursorForward } from './cursorForward'
 export { default as cursorNext } from './cursorNext'
 export { default as cursorPrev } from './cursorPrev'
 export { default as cursorUp } from './cursorUp'
+export { default as dataIntegrityCheck } from './dataIntegrityCheck'
 export { default as deleteAttribute } from './deleteAttribute'
 export { default as deleteEmptyThought } from './deleteEmptyThought'
+export { default as deleteThought } from './deleteThought'
 export { default as error } from './error'
 export { default as expandContextThought } from './expandContextThought'
 export { default as home } from './home'
+export { default as importText } from './importText'
 export { default as indent } from './indent'
 export { default as loadFromUrl } from './loadFromUrl'
 export { default as loadRemoteState } from './loadRemoteState'
@@ -33,10 +36,8 @@ export { default as toggleAttribute } from './toggleAttribute'
 export { overlayHide, scrollPrioritize, overlayReveal } from './toolbar'
 export { tutorialNext, tutorialPrev } from './tutorial'
 export { default as undoArchive } from './undoArchive'
+export { default as userAuthenticated } from './userAuthenticated'
 export { default as updateSplitPosition } from './updateSplitPosition'
-export { default as importText } from './importText'
 export { default as loadLocalState } from './loadLocalState'
 export { default as login } from './login'
 export { default as logout } from './logout'
-export { default as userAuthenticated } from './userAuthenticated'
-export { default as dataIntegrityCheck } from './dataIntegrityCheck'
