+++ src/action-creators/loadFromUrl.js
@@ -13,8 +13,6 @@ import {
 
 import { importText } from '../action-creators'
 
-import { store } from '../store'
-
 /**
  * Imports thoughts from the given source url into the given path (default: root)
  *
@@ -26,7 +24,7 @@ export default async (url, path = RANKED_ROOT, { skipRoot } = {}) => async (disp
   const text = await response.text()
 
   // prevent the default setCursor behavior of importText so that we can restore the cursor from the url
-  await store.dispatch(importText(path, text, { preventSetCursor: true, skipRoot }))
+  await dispatch(importText(path, text, { preventSetCursor: true, skipRoot }))
 
   // decode url after importText so that we are using updated state
   const state = getState()
