+++ src/util/index.js
@@ -15,7 +15,6 @@ export { conjunction } from './conjunction'
 export { contextChainToPath } from './contextChainToPath'
 export { contextOf } from './contextOf'
 export { decodeCharacterEntities } from './decodeCharacterEntities'
-export { deleteThought } from './deleteThought'
 export { download } from './download'
 export { editableNode } from './editableNode'
 export { ellipsize } from './ellipsize'
@@ -36,7 +35,7 @@ export { hashThought } from './hashThought'
 export { head } from './head'
 export { headRank } from './headRank'
 export { headValue } from './headValue'
-export { importHtml, rawTextToHtml } from './importText'
+export { importHtml } from './importHTML'
 export { initEvents } from './initEvents'
 export { initFirebase } from './initFirebase'
 export { initialState } from './initialState'
